+++ Packs/Base/Scripts/CommonServerPython/CHANGELOG.md
@@ -1,7 +1,5 @@
 ## [Unreleased]
-
-
-## [20.4.1] - 2020-04-29
+  - Added retry mechanism to the BaseClient
   - Deprecated the following enums: 
     - **entryTypes**
     - **formats**